+++ helpers/ColorHelper.php
@@ -4,7 +4,7 @@
  * Clean Theme
  * @link https://github.com/cuzy-app/clean-theme
  * @license https://github.com/cuzy-app/clean-theme/blob/master/docs/LICENCE.md
- * @author [Marc FARRE](https://marc.fun) for [CUZY.APP](https://www.cuzy.app)
+ * @author [Felix Hahn](https://github.com/felixhahnweilheim)
  */
 
 namespace humhub\modules\cleanTheme\helpers;
@@ -24,30 +24,41 @@ class ColorHelper
      */
     public static function lighten(string $color, int $amount, bool $relative = false): string
     {
-        // split color into its components
-        $colorParts = static::getColorComponents($color);
+        /*
+         * $color is expected to be a hexadecimal color code (including '#')
+         * and has to be splitted into its components
+         */
+        $color_parts = static::getColorComponents($color);
 
+        // $amount is expected to be a number between 0 an 100
         $percentage = $amount / 100;
 
         // By default the LESS lighten() function adds the $amount absolutely to L, not relatively
         if (!$relative) {
-            //Converting a RGB color to HSL, the Lightness would be calculated by L = [max(R,G,B) + min(R,G,B)] / (2 * 255)
-            $max = hexdec(max($colorParts));
-            $min = hexdec(min($colorParts));
-            if ($max != 0) {
-                $divider = (1 - ($max + $min) / (2 * 255));
-                if ($divider) {
-                    $percentage /= (1 - ($max + $min) / (2 * 255));
-                }
+
+            /*
+             * Converting a RGB color to HSL, the Lightness would be calculated by L = [max(R,G,B) + min(R,G,B)] / (2 * 255)
+             * So we need $max and $min
+             */
+            $max = hexdec(max($color_parts));
+            $min = hexdec(min($color_parts));
+
+            /*
+             * if $min is 255, we would divide by zero below
+             * and white #ffffff does not need lightening anyways
+             */
+            if ($min === 255) {
+                return $color;
             }
+            $percentage = $percentage / (1 - ($max + $min) / (2 * 255));
         }
 
         $result = '#';
 
-        foreach ($colorParts as $colorPart) {
-            $colorPart = hexdec($colorPart); // Convert to decimal
-            $colorPart = round($colorPart + (255 - $colorPart) * $percentage); // Adjust color
-            $colorPart = max(min($colorPart, 255), 0); // keep between 0 and 255
+        foreach ($color_parts as $colorPart) {
+            $colorPart   = hexdec($colorPart); // Convert to decimal
+            $colorPart   = round($colorPart + (255 - $colorPart) * $percentage); // Adjust color
+            $colorPart   = max(min($colorPart, 255), 0); // keep between 0 and 255
             $result .= str_pad(dechex($colorPart), 2, '0', STR_PAD_LEFT); // Make two char hex code
         }
 
@@ -101,10 +112,15 @@ class ColorHelper
      */
     public static function fade(string $color, int $amount): string
     {
-        $opacity = round(($amount / 100) * 255);
+        // make sure we have 6 letters code not 3
+        $color = '#' . self::getSixDigitsColor($color);
+
+        // $amount is expected to be between 0 and 100
+        $opacity = ($amount / 100) * 255;
         $opacity = max(min($opacity, 255), 0); // keep between 0 and 255
         $opacity = str_pad(dechex($opacity), 2, '0', STR_PAD_LEFT); // make 2 char hex code
 
+        // return RGBA as hex code
         return $color . $opacity;
     }
 
@@ -115,13 +131,18 @@ class ColorHelper
      */
     protected static function getColorComponents(string $color): array
     {
+        $hexstr = self::getSixDigitsColor($color);
+        return str_split($hexstr, 2);
+    }
+
+    protected static function getSixDigitsColor(string $color): string
+    {
         // Remove leading '#'
         $hexstr = ltrim($color, '#');
         // if color has just 3 digits
-        if (strlen($hexstr) === 3) {
+        if (strlen($hexstr) == 3) {
             $hexstr = $hexstr[0] . $hexstr[0] . $hexstr[1] . $hexstr[1] . $hexstr[2] . $hexstr[2];
         }
-
-        return str_split($hexstr, 2);
+        return $hexstr;
     }
 }